requirements.txt
pandas>=1.3.0
requests>=2.25.0
beautifulsoup4>=4.9.0
psycopg2-binary>=2.9.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=0.4.0
gspread>=5.0.0
pytest>=6.0.0
pytest-cov>=2.12.0
pytest-mock>=3.6.0